unittest done

+++ src/AppBundle/Controller/AddressBookController.php
@@ -109,14 +109,14 @@ class AddressBookController extends Controller
         if (!$address) {
             throw $this->createNotFoundException('The adderss does not exist');
         }
-        if ($address->getPicture()) {
+        if ($address->getPicture() && file_exists($this->getParameter('photos_directory') . '/' . $address->getPicture())) {
             unlink($this->getParameter('photos_directory') . '/' . $address->getPicture());
         }
 
         
         $em = $this->getDoctrine()->getManager();
         $em->remove($address);
-        //$em->flush();
+        $em->flush();
         $this->addFlash(
             'notice',
             'Address successfully deleted!'
